api/requirements.txt
aiosqlite
APScheduler
asgiref
beautifulsoup4
cryptography
flake8
flask
flask>=2.0
Flask[async]
google-genai
googlesearch-python
gunicorn
httpx
pydantic>=2.0
pytest
pytest-asyncio
python-dotenv
python-telegram-bot>=21.0
respx
structlog
